Migrate IntEnumLiteral into core.
