+++ numba_dpex/core/types/__init__.py
@@ -5,6 +5,7 @@
 from .array_type import Array
 from .dpctl_types import DpctlSyclEvent, DpctlSyclQueue
 from .dpnp_ndarray_type import DpnpNdArray
+from .kernel_api.literal_intenum import IntEnumLiteral
 from .kernel_api.ranges import NdRangeType, RangeType
 from .numba_types_short_names import (
     b1,
@@ -36,8 +37,9 @@ __all__ = [
     "DpctlSyclQueue",
     "DpctlSyclEvent",
     "DpnpNdArray",
-    "RangeType",
+    "IntEnumLiteral",
     "NdRangeType",
+    "RangeType",
     "USMNdArray",
     "none",
     "boolean",
@@ -57,6 +59,6 @@ __all__ = [
     "f8",
     "float_",
     "double",
-    "void",
     "usm_ndarray",
+    "void",
 ]
