+++ numba_dpex/experimental/models.py
@@ -6,10 +6,9 @@
 numba_dpex.experimental module.
 """
 
-from llvmlite import ir as llvmir
 from numba.core import types
 from numba.core.datamodel import DataModelManager, models
-from numba.core.datamodel.models import PrimitiveModel, StructModel
+from numba.core.datamodel.models import StructModel
 from numba.core.extending import register_model
 
 import numba_dpex.core.datamodel.models as dpex_core_models
@@ -20,7 +19,6 @@ from numba_dpex.experimental.core.types.kernel_api.items import (
 )
 
 from .dpcpp_types import AtomicRefType
-from .literal_intenum_type import IntEnumLiteral
 from .types import KernelDispatcherType
 
 
@@ -35,17 +33,6 @@ class AtomicRefModel(StructModel):
             ),
         ]
         super().__init__(dmm, fe_type, members)
-
-
-class IntEnumLiteralModel(PrimitiveModel):
-    """Representation of an object of LiteralIntEnum type using Numba's
-    PrimitiveModel that can be represented natively in the target in all
-    usage contexts.
-    """
-
-    def __init__(self, dmm, fe_type):
-        be_type = llvmir.IntType(fe_type.bitwidth)
-        super().__init__(dmm, fe_type, be_type)
 
 
 class EmptyStructModel(StructModel):
@@ -71,7 +58,6 @@ def _init_exp_data_model_manager() -> DataModelManager:
     dmm = dpex_core_models.dpex_data_model_manager.copy()
 
     # Register the types and data model in the DpexExpTargetContext
-    dmm.register(IntEnumLiteral, IntEnumLiteralModel)
     dmm.register(AtomicRefType, AtomicRefModel)
 
     # Register the GroupType type
