+++ numba_dpex/experimental/target.py
@@ -8,11 +8,8 @@ eventually move into the numba_dpex.core.
 
 from functools import cached_property
 
-from llvmlite import ir as llvmir
-from numba.core import types
 from numba.core.descriptors import TargetDescriptor
 from numba.core.target_extension import GPU, target_registry
-from numba.core.types.scalars import IntEnumClass
 
 from numba_dpex._kernel_api_impl.spirv.target import (
     SPIRVTargetContext,
@@ -20,9 +17,6 @@ from numba_dpex._kernel_api_impl.spirv.target import (
 )
 from numba_dpex.core.descriptor import DpexTargetOptions
 from numba_dpex.experimental.models import exp_dmm
-from numba_dpex.kernel_api.flag_enum import FlagEnum
-
-from .literal_intenum_type import IntEnumLiteral
 
 
 #  pylint: disable=R0903
@@ -45,37 +39,6 @@ class DpexExpKernelTypingContext(SPIRVTypingContext):
     are stable enough to be migrated to DpexKernelTypingContext.
     """
 
-    def resolve_value_type(self, val):
-        """
-        Return the numba type of a Python value that is being used
-        as a runtime constant.
-        ValueError is raised for unsupported types.
-        """
-
-        typ = super().resolve_value_type(val)
-
-        if isinstance(typ, IntEnumClass) and issubclass(val, FlagEnum):
-            typ = IntEnumLiteral(val)
-
-        return typ
-
-    def resolve_getattr(self, typ, attr):
-        """
-        Resolve getting the attribute *attr* (a string) on the Numba type.
-        The attribute's type is returned, or None if resolution failed.
-        """
-        retty = None
-
-        if isinstance(typ, IntEnumLiteral):
-            try:
-                attrval = getattr(typ.literal_value, attr).value
-                retty = types.IntegerLiteral(attrval)
-            except ValueError:
-                pass
-        else:
-            retty = super().resolve_getattr(typ, attr)
-        return retty
-
 
 #  pylint: disable=W0223
 # FIXME: Remove the pylint disablement once we add an override for
@@ -94,22 +57,6 @@ class DpexExpKernelTargetContext(SPIRVTargetContext):
     def __init__(self, typingctx, target=DPEX_KERNEL_EXP_TARGET_NAME):
         super().__init__(typingctx, target)
         self.data_model_manager = exp_dmm
-
-    def get_getattr(self, typ, attr):
-        """
-        Overrides the get_getattr function to provide an implementation for
-        getattr call on an IntegerEnumLiteral type.
-        """
-
-        if isinstance(typ, IntEnumLiteral):
-            #  pylint: disable=W0613
-            def enum_literal_getattr_imp(context, builder, typ, val, attr):
-                enum_attr_value = getattr(typ.literal_value, attr).value
-                return llvmir.Constant(llvmir.IntType(64), enum_attr_value)
-
-            return enum_literal_getattr_imp
-
-        return super().get_getattr(typ, attr)
 
 
 class DpexExpKernelTarget(TargetDescriptor):
