+++ numba_dpex/tests/core/types/IntEnumLiteral/test_compilation.py
